Errata
> ## Office of the State Registrar
> ### July 6, 2025

**ERRATA**

SB269, the "DPS Dissolution Act" originally had Section II(b) reading as:
> 3 M.S.C. 1 § 2208 and all subordinate clauses and subsections shall be repealed in full.

The Senate, during their session, passed something stating the following:
> SA002; amending section IIb to state "3 M.S.C. 1 sec. 2208(a) shall be repealed in full." agreed to without objection

Neither the PDF or the Google Document were updated, and no notice was given to the team in charge of the Mayflower State Code. This resulted in 3 M.S.C. 1 §§ 2208—2208(b) being marked as "repealed" on the State Code.